MPT-16437 Reorganise resource commerce mixins structure
