+++ pyproject.toml
@@ -140,7 +140,6 @@ per-file-ignores = [
   "tests/unit/resources/accounts/*.py: WPS204 WPS202 WPS210",
   "tests/unit/resources/catalog/test_products.py: WPS202 WPS210",
   "tests/unit/resources/commerce/*.py: WPS202 WPS204",
-  "tests/unit/resources/*/test_mixins.py: WPS118 WPS202 WPS204 WPS235",
   "tests/unit/test_mpt_client.py: WPS235",
   "tests/*: WPS432 WPS202",
 ]
